Removal of the DockerService
